Improve module structure.
