+++ packages/editor/src/client/blocks/richTextBlock/buttons.tsx
@@ -1,7 +1,4 @@
-import React, {useState} from 'react'
-import {useTranslation} from 'react-i18next'
-import {Button, ControlLabel, Form, FormControl, FormGroup, Modal} from 'rsuite'
-import {Editor, Transforms, Range} from 'slate'
+import React from 'react'
 import {useSlate} from 'slate-react'
 import {
   ToolbarIconButtonProps,
@@ -10,13 +7,13 @@ import {
   ToolbarIconButton
 } from '../../atoms/toolbar'
 import {isFormatActive, toggleFormat} from './editorUtils'
-import {Format, InlineFormat} from './formats'
+import {Format} from './formats'
 
-interface SlateBlockButtonProps extends ToolbarIconButtonProps {
+interface SlateBlockIconButtonProps extends ToolbarIconButtonProps {
   readonly format: Format
 }
 
-export function FormatIconButton({icon, format}: SlateBlockButtonProps) {
+export function FormatIconButton({icon, format}: SlateBlockIconButtonProps) {
   const editor = useSlate()
 
   return (
@@ -31,11 +28,11 @@ export function FormatIconButton({icon, format}: SlateBlockButtonProps) {
   )
 }
 
-interface SlateBlockButtonWithChildrenProps extends ToolbarButtonProps {
+interface SlateBlockButtonProps extends ToolbarButtonProps {
   readonly format: Format
 }
 
-export function FormatButtonWithChildren({format, children}: SlateBlockButtonWithChildrenProps) {
+export function FormatButton({format, children}: SlateBlockButtonProps) {
   const editor = useSlate()
 
   return (
@@ -48,144 +45,6 @@ export function FormatButtonWithChildren({format, children}: SlateBlockButtonWit
       {children}
     </ToolbarButton>
   )
-}
-
-export function LinkFormatButton() {
-  const editor = useSlate()
-  const [isLinkDialogOpen, setLinkDialogOpen] = useState(false)
-  const [selection, setSelection] = useState<Range | null>(null)
-
-  const [title, setTitle] = useState('')
-  const [url, setURL] = useState('')
-
-  const validatedURL = validateURL(url)
-  const isDisabled = !validatedURL
-
-  const {t} = useTranslation()
-
-  return (
-    <>
-      <ToolbarIconButton
-        icon="link"
-        active={isFormatActive(editor, InlineFormat.Link)}
-        onMouseDown={e => {
-          e.preventDefault()
-
-          const nodes = Array.from(
-            Editor.nodes(editor, {
-              at: editor.selection ?? undefined,
-              match: node => node.type === InlineFormat.Link
-            })
-          )
-
-          const tuple = nodes[0]
-
-          if (tuple) {
-            const [node] = tuple
-
-            setTitle((node.title as string) ?? '')
-            setURL((node.url as string) ?? '')
-          } else {
-            setTitle('')
-            setURL('')
-          }
-
-          setSelection(editor.selection)
-          setLinkDialogOpen(true)
-        }}
-      />
-      <Modal show={isLinkDialogOpen} size={'sm'} onHide={() => setLinkDialogOpen(false)}>
-        <Modal.Body>
-          <Form fluid={true}>
-            <FormGroup>
-              <ControlLabel>{t('blocks.richText.link')}</ControlLabel>
-              <FormControl
-                errorMessage={url && !validatedURL ? 'Invalid Link' : undefined}
-                value={url}
-                onChange={url => setURL(url)}
-              />
-            </FormGroup>
-            <FormGroup>
-              <ControlLabel>{t('blocks.richText.title')}</ControlLabel>
-              <FormControl value={title} onChange={title => setTitle(title)} />
-            </FormGroup>
-          </Form>
-        </Modal.Body>
-        <Modal.Footer>
-          <Button
-            disabled={isDisabled}
-            onClick={() => {
-              insertLink(editor, selection, validatedURL!, title || undefined)
-              setLinkDialogOpen(false)
-            }}>
-            {t('Apply')}
-          </Button>
-          <Button onClick={() => setLinkDialogOpen(false)}>{t('Cancel')}</Button>
-        </Modal.Footer>
-      </Modal>
-    </>
-  )
-}
-
-function validateURL(url: string): string | null {
-  return url
-
-  // TODO: Implement better URL validation with for support relative links.
-  // try {
-  //   return new URL(url).toString()
-  // } catch (err) {
-  //   try {
-  //     return new URL(`https://${url}`).toString()
-  //   } catch (err) {
-  //     return null
-  //   }
-  // }
-}
-
-export function RemoveLinkFormatButton() {
-  const editor = useSlate()
-
-  return (
-    <ToolbarIconButton
-      icon="unlink"
-      active={isFormatActive(editor, InlineFormat.Link)}
-      disabled={!isFormatActive(editor, InlineFormat.Link)}
-      onMouseDown={e => {
-        e.preventDefault()
-        removeLink(editor)
-      }}
-    />
-  )
-}
-
-function insertLink(editor: Editor, selection: Range | null, url: string, title?: string) {
-  if (selection) {
-    if (Range.isCollapsed(selection)) {
-      const nodes = Array.from(
-        Editor.nodes(editor, {
-          at: selection,
-          match: node => node.type === InlineFormat.Link
-        })
-      )
-
-      const tuple = nodes[0]
-
-      if (tuple) {
-        const [, path] = tuple
-        Transforms.select(editor, path)
-      }
-    } else {
-      Transforms.select(editor, selection)
-    }
-  }
-
-  Transforms.unwrapNodes(editor, {match: node => node.type === InlineFormat.Link})
-  Transforms.wrapNodes(editor, {type: InlineFormat.Link, url, title, children: []}, {split: true})
-  Transforms.collapse(editor, {edge: 'end'})
-}
-
-function removeLink(editor: Editor) {
-  Transforms.unwrapNodes(editor, {match: node => node.type === InlineFormat.Link})
 }
 
 export function H1Icon() {
